Report pdf
Hey I started on the report just a bit. Its at the bottom of the PDF. I basically just tried to summarize how Ty processed the data. I still need to add a description of the data along with the stuff about what the algorithm produces.. Im going to work on it tomorrow morning a lot :)